+++ app/src/main/java/com/feitianzhu/huangliwo/common/base/activity/AbsActivity.java
@@ -36,7 +36,6 @@ public abstract class AbsActivity extends AppCompatActivity {
         }
 
     }
-
     public void initBase() {
 
     }
