+++ app/src/main/java/com/feitianzhu/huangliwo/shop/Customerservice.java
@@ -86,37 +86,16 @@ public class Customerservice extends BaseActivity {
                 finish();
                 break;
             case R.id.shopissues:
+
                 LoadingUtil.setLoadingViewShow(false);
-                if (response != null && response.size() >= 0) {
-                    EMClient.getInstance().login("14701776629", "123456", new EMCallBack() {//回调
-                        @Override
-                        public void onSuccess() {
-                            EMClient.getInstance().groupManager().loadAllGroups();
-                            EMClient.getInstance().chatManager().loadAllConversations();
-                            //startActivity(new Intent(Customerservice.this,ImActivity.class));
-                            Intent intent = new Intent(Customerservice.this, ImActivity.class);
-                            //username为对方的环信id
-//                            intent.putExtra(EaseConstant.EXTRA_USER_ID, response.get(0).getUserId());
-                            intent.putExtra(EaseConstant.EXTRA_USER_ID, "13671192850");
-                            startActivity(intent);
-                            Log.d("main", "登录聊天服务器成功!");
-                            LoadingUtil.setLoadingViewShow(false);
-                        }
-
-                        @Override
-                        public void onProgress(int progress, String status) {
-
-                        }
-
-                        @Override
-                        public void onError(int code, String message) {
-                            LoadingUtil.setLoadingViewShow(false);
-                            Log.i("onError", "onError: " + code + message);
-                            Log.d("main", "登录聊天服务器失败！");
-                        }
-                    });
+                if (response!=null &&response.size()>0) {
+                    Intent intent = new Intent(Customerservice.this, ImActivity.class);
+                    //username为对方的环信id
+                    intent.putExtra("name",response.get(0).getNick());
+                    intent.putExtra("icon",response.get(0).getIcon());
+                    intent.putExtra(EaseConstant.EXTRA_USER_ID, response.get(0).getUserId()+"-dev");
+                    startActivity(intent);
                 }
-
                 break;
             case R.id.returnissue:
                 startActivity(new Intent(Customerservice.this, CustomerservicelistActivity.class));
