+++ app/src/main/res/values/colors.xml
@@ -68,6 +68,7 @@
     <color name="color_000000">#000000</color>
     <color name="color_fed228">#fed228</color>
     <color name="color_E63411">#E63411</color>
+    <color name="color_2A2420">#2A2420</color>
     <color name="color_CCCCCC">#CCCCCC</color>
     <color name="color_green">#13D1BE</color>
     <color name="color_ff8300">#ff8300</color>
